Factored out Not
