+++ tests/basics.rs
@@ -85,6 +85,7 @@ async fn main() {
     BasicsWorld::run("features/validation/booleans.feature").await;
     BasicsWorld::run("features/validation/const.feature").await;
     BasicsWorld::run("features/validation/enums.feature").await;
+    BasicsWorld::run("features/validation/not.feature").await;
     BasicsWorld::run("features/validation/nulls.feature").await;
     BasicsWorld::run("features/validation/numbers.feature").await;
     BasicsWorld::run("features/validation/objects.feature").await;
